Update packages in project gen + rename to compose
