+++ platforms/intellij/compose/build.gradle.kts
