+++ platforms/intellij/compose/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=skate-compose
+POM_NAME=Skate (Compose)
+POM_DESCRIPTION=Skate (Compose)
+INTELLIJ_PLUGIN=true
