--- platforms/intellij/project-gen/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=skate-project-gen
-POM_NAME=Skate (Project Generator)
-POM_DESCRIPTION=Skate (Project Generator)
-INTELLIJ_PLUGIN=true
