--- platforms/intellij/project-gen/src/jvmMain/kotlin/slack/tooling/projectgen/SlackDesktopTheme.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright (C) 2024 Slack Technologies, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package slack.tooling.projectgen
-
-import androidx.compose.runtime.Composable
-import org.jetbrains.jewel.bridge.theme.SwingBridgeTheme
-import org.jetbrains.jewel.foundation.ExperimentalJewelApi
-
-@OptIn(ExperimentalJewelApi::class)
-@Composable
-fun SlackDesktopTheme(content: @Composable () -> Unit) {
-  SwingBridgeTheme(content = content)
-}
