+++ settings.gradle.kts
@@ -150,8 +150,8 @@ include(
   ":platforms:gradle:agp-handlers:agp-handler-api",
   ":platforms:gradle:slack-plugin",
   ":platforms:intellij:artifactory-authenticator",
-  ":platforms:intellij:compose-playground",
-  ":platforms:intellij:project-gen",
+  ":platforms:intellij:compose",
+  ":platforms:intellij:compose:playground",
   ":platforms:intellij:skate",
   ":tools:foundry-common",
   ":tools:skippy",
